fix(review): move executable binding behavior to runtime
